#193: Rename file to contactDetails
